refactor: 로직 훅으로 분리
